pyproject.toml
[project]
name = "resolve-ai"
version = "0.1.0"
description = "Multi-agent chatbot for Brazilian consumer rights (CDC) — RAG + LangGraph + Gemini + FastAPI"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "google-genai>=1.0",
    "httpx>=0.28.1",
    "langchain>=0.3",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "llama-index>=0.14.15",
    "pydantic>=2",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
